+++ docs/server/generate_http_server_api_documentation.py
@@ -5,7 +5,7 @@ import os
 import os.path
 
 from bigchaindb.common.transaction import Transaction, Input, TransactionLink
-from bigchaindb.core import Bigchain
+from bigchaindb.tendermint import BigchainDB
 from bigchaindb.tendermint import lib
 from bigchaindb.web import server
