+++ tests/test_core.py
@@ -26,10 +26,10 @@ def config(request, monkeypatch):
 
 
 def test_bigchain_class_default_initialization(config):
-    from bigchaindb.core import Bigchain
+    from bigchaindb.tendermint import BigchainDB
     from bigchaindb.consensus import BaseConsensusRules
     from bigchaindb.backend.connection import Connection
-    bigchain = Bigchain()
+    bigchain = BigchainDB()
     assert isinstance(bigchain.connection, Connection)
     assert bigchain.connection.host == config['database']['host']
     assert bigchain.connection.port == config['database']['port']
@@ -38,7 +38,7 @@ def test_bigchain_class_default_initialization(config):
 
 
 def test_bigchain_class_initialization_with_parameters(config):
-    from bigchaindb.core import Bigchain
+    from bigchaindb.tendermint import BigchainDB
     from bigchaindb.backend import connect
     from bigchaindb.consensus import BaseConsensusRules
     init_db_kwargs = {
@@ -48,7 +48,7 @@ def test_bigchain_class_initialization_with_parameters(config):
         'name': 'this_is_the_db_name',
     }
     connection = connect(**init_db_kwargs)
-    bigchain = Bigchain(connection=connection, **init_db_kwargs)
+    bigchain = BigchainDB(connection=connection, **init_db_kwargs)
     assert bigchain.connection == connection
     assert bigchain.connection.host == init_db_kwargs['host']
     assert bigchain.connection.port == init_db_kwargs['port']
@@ -58,13 +58,13 @@ def test_bigchain_class_initialization_with_parameters(config):
 
 def test_get_blocks_status_containing_tx(monkeypatch):
     from bigchaindb.backend import query as backend_query
-    from bigchaindb.core import Bigchain
+    from bigchaindb.tendermint import BigchainDB
     blocks = [
         {'id': 1}, {'id': 2}
     ]
     monkeypatch.setattr(backend_query, 'get_blocks_status_from_transaction', lambda x: blocks)
-    monkeypatch.setattr(Bigchain, 'block_election_status', lambda x, y, z: Bigchain.BLOCK_VALID)
-    bigchain = Bigchain(public_key='pubkey', private_key='privkey')
+    monkeypatch.setattr(BigchainDB, 'block_election_status', lambda x, y, z: BigchainDB.BLOCK_VALID)
+    bigchain = BigchainDB(public_key='pubkey', private_key='privkey')
     with pytest.raises(Exception):
         bigchain.get_blocks_status_containing_tx('txid')
